todo_web/backend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlmodel>=0.0.22
psycopg2-binary>=2.9.9
asyncpg>=0.30.0

# Authentication
httpx>=0.27.0  # For Better Auth session validation
python-jose[cryptography]>=3.3.0
PyJWT>=2.9.0
bcrypt>=4.2.0

# Rate limiting
slowapi>=0.1.9

# Utilities
python-dotenv>=1.0.1
pydantic[email]>=2.9.0
pydantic-settings>=2.6.0

# CORS
python-multipart>=0.0.12

# Date handling
python-dateutil>=2.9.0
